refactor: extract applyWidth into AppliesSectionWidth trait
SectionComponentFactory and SectionInfolistsFactory carried a byte-identical
private applyWidth() method. Extract it into a shared trait so there is a
single source of truth for the width-application rule.
